requirements.txt
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
blessings==1.6
decorator==4.0.10
Django==1.9.7
django-extensions==1.6.7
django-nose==1.4.4
django-test-without-migrations==0.4
djangorestframework==3.3.3
factory-boy==2.7.0
fake-factory==0.5.9
ipaddress==1.0.16
ipython==5.0.0
ipython-genutils==0.1.0
nose==1.3.7
nose-progressive==1.5.1
pathlib2==2.1.0
pexpect==4.2.0
pickleshare==0.7.2
prompt-toolkit==1.0.3
psycopg2==2.6.1
ptyprocess==0.5.1
Pygments==2.1.3
python-dateutil==2.5.3
simplegeneric==0.8.1
six==1.10.0
traitlets==4.2.2
wcwidth==0.1.7
Werkzeug==0.11.10
